sound: add pause/unpause/stop to libtrx
