+++ src/libtrx/include/libtrx/game/sound/common.h
@@ -4,5 +4,10 @@
 #include "enum.h"
 #include "ids.h"
 
-bool Sound_Effect(SOUND_EFFECT_ID sfx_num, const XYZ_32 *pos, uint32_t flags);
-bool Sound_IsAvailable(SOUND_EFFECT_ID sfx_num);
+void Sound_PauseAll(void);
+void Sound_UnpauseAll(void);
+
+extern void Sound_StopAll(void);
+extern bool Sound_Effect(
+    SOUND_EFFECT_ID sfx_num, const XYZ_32 *pos, uint32_t flags);
+extern bool Sound_IsAvailable(SOUND_EFFECT_ID sfx_num);
