+++ src/tr1/game/sound.h
@@ -2,20 +2,17 @@
 
 #include "global/types.h"
 
-#include <stdbool.h>
+#include <libtrx/game/sound.h>
+
 #include <stddef.h>
 #include <stdint.h>
 
 bool Sound_Init(void);
 void Sound_Shutdown(void);
-bool Sound_Effect(SOUND_EFFECT_ID sfx_num, const XYZ_32 *pos, uint32_t flags);
 bool Sound_StopEffect(SOUND_EFFECT_ID sfx_num, const XYZ_32 *pos);
 void Sound_UpdateEffects(void);
 void Sound_ResetEffects(void);
 void Sound_StopAmbientSounds(void);
-void Sound_PauseAll(void);
-void Sound_UnpauseAll(void);
-void Sound_StopAllSamples(void);
 int8_t Sound_GetMasterVolume(void);
 void Sound_SetMasterVolume(int8_t volume);
 int32_t Sound_GetMinVolume(void);
