+++ src/tr2/game/phase/phase_game.c
@@ -63,7 +63,7 @@ static void M_End(PHASE *const phase)
     Game_SetIsPlaying(false);
     M_PRIV *const p = phase->priv;
     Overlay_HideGameInfo();
-    Sound_StopAllSamples();
+    Sound_StopAll();
     Music_Stop();
     Music_SetVolume(g_Config.audio.music_volume);
 }
